third_party/blink/web_tests/http/tests/contacts/select-function.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources.js"></script>

<script>
promise_test(function() {
  return assert_promise_rejects_with_message(
      navigator.contacts.select({multiple: true, properties: ['name', 'email', 'tel']}),
      new DOMException('Unable to open a contact selector', 'AbortError'),
      'navigator.contacts.select');
}, 'Select function throws exception.');
</script>
